+++ src/cacheServerClient/cacheServerClient.types.ts
@@ -92,6 +92,7 @@ export interface ClaimData extends Record<string, unknown> {
     profile?: Profile;
     claimType?: string;
     claimTypeVersion?: number;
+    claimParams?: Record<string, string>;
 }
 
 export enum Order {